* Spectrogram RGB transform should return float32 not float64

+++ data_interfaces/transforms.py
@@ -167,7 +167,7 @@ class ApplyColormapToSpectrogram(object):
         if 'cuda' in original_device.type:
             img = img.cpu()
 
-        sample['image'] = torch.tensor(self.cmap(img, alpha=None)[:, :, :3], device=original_device).permute(2, 0, 1)
+        sample['image'] = torch.tensor(self.cmap(img, alpha=None)[:, :, :3], device=original_device, dtype=torch.float32).permute(2, 0, 1)
 
         return sample
